move from/to fen
